split modules into files. reraise if validation fails
